refactor(rust): move HTML lexer skeleton into Rust package
