+++ code/packages/rust/html-lexer/BUILD_windows
@@ -1,0 +1,1 @@
+cargo test -p coding-adventures-html-lexer -- --nocapture
