+++ code/packages/rust/html-lexer/CHANGELOG.md
@@ -1,0 +1,12 @@
+# Changelog
+
+All notable changes to the `coding-adventures-html-lexer` crate will be
+documented in this file.
+
+## [0.1.0] - 2026-04-23
+
+### Added
+- Initial Rust HTML lexer package.
+- Statically linked HTML skeleton state machine over the generic
+  `state-machine-tokenizer` runtime.
+- Build-time `html-skeleton.lexer.states.toml` authoring artifact.
